Move pagination stuff Fix some docs
