+++ src/main/java/com/freya02/botcommands/api/pagination/menu/ChoiceMenu.java
@@ -1,13 +1,13 @@
-package com.freya02.botcommands.api.pagination.menu2;
+package com.freya02.botcommands.api.pagination.menu;
 
 import com.freya02.botcommands.api.components.Components;
 import com.freya02.botcommands.api.components.event.ButtonEvent;
-import com.freya02.botcommands.api.pagination.Paginator;
+import com.freya02.botcommands.api.pagination.ButtonContentSupplier;
 import com.freya02.botcommands.api.pagination.PaginatorSupplier;
 import com.freya02.botcommands.api.pagination.TimeoutInfo;
-import com.freya02.botcommands.api.pagination.menu.ButtonContent;
-import com.freya02.botcommands.api.pagination.menu.ButtonContentSupplier;
+import com.freya02.botcommands.api.pagination.paginator.Paginator;
 import com.freya02.botcommands.api.pagination.transformer.EntryTransformer;
+import com.freya02.botcommands.api.utils.ButtonContent;
 import net.dv8tion.jda.api.interactions.components.Button;
 
 import java.util.List;
