+++ src/main/java/com/freya02/botcommands/api/pagination/menu/RowPrefixSupplier.java
@@ -1,4 +1,4 @@
-package com.freya02.botcommands.api.pagination.menu2;
+package com.freya02.botcommands.api.pagination.menu;
 
 public interface RowPrefixSupplier {
 	String apply(int entryNum, int maxEntries);
