+++ src/main/java/com/freya02/botcommands/api/pagination/paginator/Paginator.java
@@ -1,8 +1,11 @@
-package com.freya02.botcommands.api.pagination;
+package com.freya02.botcommands.api.pagination.paginator;
 
 import com.freya02.botcommands.api.BContext;
-import com.freya02.botcommands.api.pagination.menu.ButtonContent;
-import com.freya02.botcommands.api.pagination.menu2.Menu;
+import com.freya02.botcommands.api.pagination.PaginatorComponents;
+import com.freya02.botcommands.api.pagination.PaginatorSupplier;
+import com.freya02.botcommands.api.pagination.TimeoutInfo;
+import com.freya02.botcommands.api.pagination.menu.Menu;
+import com.freya02.botcommands.api.utils.ButtonContent;
 
 /**
  * Provides a paginator
